Add "Shading" mode
In this case, when the user enters the grid for the first time,
that grid cell will be filled with the lightest spectrum of black color.
Then, every time the user enters the cell again,
the color of that cell becomes 10% bolder.
until it finally reaches the boldest possible state of black.